+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+1.0.0-rc.27
+=============
+* Added wait to form element before filling a value
+* Display variation name in PHPUnit test result log file
+
 1.0.0-rc.26
 =============
 * Add an ability to specify several values for one tag type